+++ src/GameDisplayGui.cpp
@@ -19,7 +19,7 @@ void GameDisplay::drawStat(int x, int y, String texture, long val, int animTick)
     t1.setScale(animTick / 30.f + 1.f, animTick / 30.f + 1.f);
     this->renderWnd->draw(t1);
 
-    if(sf::IntRect(x, y, 50, 50).contains(Mouse::getPosition(*renderWnd)) || Game::instance->isNewPlayer)
+    if(sf::IntRect(x, y, 50, 50).contains(mousePos()) || Game::instance->isNewPlayer)
     {
         t1.move(0.f, 30.f);
         t1.setOrigin(0.f, 0.f);
@@ -92,7 +92,7 @@ void GameDisplay::setSplash(String text)
 
 Vector2i GameDisplay::mousePos()
 {
-	return Vector2i(this->getRenderWnd()->mapPixelToCoords(Mouse::getPosition(*this->getRenderWnd())));
+	return Vector2i(this->getRenderWnd()->mapPixelToCoords(Mouse::getPosition(*this->getRenderWnd()), Game::instance->getGameView()));
 }
 
 void GameDisplay::drawDebugPie(sf::RenderWindow* wnd)
@@ -257,10 +257,8 @@ void GameDisplay::drawGui()
         drawStat(850, 32, "points_mpl", game->pointsToNewMpl);
     }
 
-	if(game->isGuiLoaded())
-        game->drawGui(false);
-
     drawStat(50, 32, "coin", game->getCoins());
+	game->drawGui(false);
 
     if(game->debug) drawDebugInfo(this->renderWnd);
 }