+++ src/GuiPowers.cpp
@@ -66,21 +66,21 @@ void GuiPowers::onLoad()
         addWidget(&epd.bImg);
     }
 
-    addWidget(&(bReturn = Button(this, Vector2f(400.f, 40.f), Vector2f(game->getSize().x / 2 - 200, game->getSize().y * 3 / 4),
+    addWidget(&(bReturn = Button(this, Vector2f(400.f, 40.f), Vector2f(0, 0),
                                Game::instance->translation.get("gui.powers.return"), 0)));
 }
 void GuiPowers::onResize()
 {
     GameDisplay* game = GameDisplay::instance;
-    bReturn.setPosition(Vector2f(game->getSize().x / 2 - 200, game->getSize().y * 3 / 4));
+    bReturn.setPosition(Vector2f(game->getSize().x / 2 - 200, powerData.size() * 50.f + 470.f));
     for(PowerData* data: powerData)
     {
-        data->bBuyPower.setPosition(Vector2f(game->getSize().x / 2 - 300.f, (data->bBuyPower.getID() - 100) * 50.f + game->getSize().y / 4));
-        data->bEquipPower.setPosition(Vector2f(game->getSize().x / 2 + 260.f, (data->bBuyPower.getID() - 100) * 50.f + game->getSize().y / 4));
+        data->bBuyPower.setPosition(Vector2f(game->getSize().x / 2 - 300.f, (data->bBuyPower.getID() - 100) * 50.f + 200.f));
+        data->bEquipPower.setPosition(Vector2f(game->getSize().x / 2 + 260.f, (data->bBuyPower.getID() - 100) * 50.f + 200.f));
     }
     for(size_t s = 0; s < equippedPowerIds.size(); s++)
     {
-        equippedPowerIds[s].bImg.setPosition(Vector2f(game->getSize().x / 2 - 45.f + s * 50.f, powerData.size() * 50.f + game->getSize().y / 4 + 100.f));
+        equippedPowerIds[s].bImg.setPosition(Vector2f(game->getSize().x / 2 - 45.f + s * 50.f, powerData.size() * 50.f + 100.f + 200.f));
     }
 }
 GuiPowers::~GuiPowers()
@@ -129,7 +129,7 @@ void GuiPowers::onDraw(sf::RenderWindow& wnd)
     }
 
     // equipped powers
-    wnd.draw(GameDisplay::instance->drawCenteredString(Game::instance->translation.get("gui.powers.equippedPowers"), 20, Vector2f(game->getSize().x / 2, powerData.size() * 50.f + game->getSize().y / 4 + 75.f)));
+    wnd.draw(GameDisplay::instance->drawCenteredString(Game::instance->translation.get("gui.powers.equippedPowers"), 20, Vector2f(game->getSize().x / 2, powerData.size() * 50.f + 275.f)));
 
     for(size_t s = 0; s < equippedPowerIds.size(); s++)
     {
@@ -153,9 +153,9 @@ void GuiPowers::onDraw(sf::RenderWindow& wnd)
         bool bEquipClicked=data->bEquipPower.isClicked(scaledPos), bBuyClicked=data->bBuyPower.isClicked(scaledPos);
         if(bEquipClicked || bBuyClicked)
         {
-            wnd.draw(drawCenteredString(Game::instance->translation.get("power." + data->power->getName() + ".desc"), 20, Vector2f(Game::instance->getSize().x / 2, powerData.size() * 50.f + Game::instance->getSize().y / 4 + 160.f)));
-            wnd.draw(drawCenteredString(Game::instance->translation.get("power." + data->power->getName() + ".upgr"), 20, Vector2f(Game::instance->getSize().x / 2, powerData.size() * 50.f + Game::instance->getSize().y / 4 + 200.f)));
-            wnd.draw(drawCenteredString(Game::instance->translation.get("gui.powers.equip"), 20, Vector2f(Game::instance->getSize().x / 2, powerData.size() * 50.f + Game::instance->getSize().y / 4 + 240.f), Text::Bold));
+            wnd.draw(drawCenteredString(Game::instance->translation.get("power." + data->power->getName() + ".desc"), 20, Vector2f(Game::instance->getSize().x / 2, powerData.size() * 50.f + 360.f)));
+            wnd.draw(drawCenteredString(Game::instance->translation.get("power." + data->power->getName() + ".upgr"), 20, Vector2f(Game::instance->getSize().x / 2, powerData.size() * 50.f + 400.f)));
+            wnd.draw(drawCenteredString(Game::instance->translation.get("gui.powers.equip"), 20, Vector2f(Game::instance->getSize().x / 2, powerData.size() * 50.f + 440.f), Text::Bold));
             break; //only one string at once can be displayed.
         }
     }
@@ -224,7 +224,7 @@ void GuiPowers::onClick(int button)
                 game->equippedPowers[s] = powerId;
                 game->savePlayerData();
                 equippedPowerIds[s].bImg = ButtonImage(this, "power/" + to_string(powerId), Vector2f(40.f, 40.f), Vector2f(), "", equippedPowerIds[s].bImg.getID());
-                equippedPowerIds[s].bImg.setPosition(Vector2f(game->getSize().x / 2 - 45.f + s * 50.f, powerData.size() * 50.f + game->getSize().y / 4 + 100.f));
+                equippedPowerIds[s].bImg.setPosition(Vector2f(game->getSize().x / 2 - 45.f + s * 50.f, powerData.size() * 50.f + 300.f));
                 break;
             }
         }
@@ -234,7 +234,7 @@ void GuiPowers::onClick(int button)
         int slotId = button - 300;
         equippedPowerIds[slotId].powerId = 0;
         equippedPowerIds[slotId].bImg = ButtonImage(this, "stat/mpl", Vector2f(40.f, 40.f), Vector2f(), "", equippedPowerIds[slotId].bImg.getID());
-        equippedPowerIds[slotId].bImg.setPosition(Vector2f(game->getSize().x / 2 - 45.f + slotId * 50.f, powerData.size() * 50.f + game->getSize().y / 4 + 100.f));
+        equippedPowerIds[slotId].bImg.setPosition(Vector2f(game->getSize().x / 2 - 45.f + slotId * 50.f, powerData.size() * 50.f + 300.f));
         equippedPowerIds[slotId].bImg.setEnabled(false);
         game->equippedPowers[slotId] = 0;
         game->savePlayerData();