+++ src/Level.cpp
@@ -4,7 +4,7 @@ namespace LevelUtility
 {
 	int getLaneFromPos(Vector2f position)
 	{
-		float windowMiddle = GameDisplay::instance->getRenderWnd()->getSize().y / 2;
+		float windowMiddle = 540.f;
 		if (position.y <= windowMiddle - 50)
 			return 0;
 		else if (position.y > windowMiddle - 50 && position.y < windowMiddle + 50)
@@ -16,7 +16,7 @@ namespace LevelUtility
 
 	Vector2f getLanePos(int lane)
 	{
-		//	                |--------------------window center---------------------|
-		return Vector2f(0.f, (GameDisplay::instance->getRenderWnd()->getSize().y / 2) + (lane-1) * 100);
+		//	            |-window center-|
+		return Vector2f(0.f, 540.f + (lane-1) * 100);
 	}
 }