+++ src/PowerFreeze.cpp
@@ -23,8 +23,7 @@ void PowerFreeze::drawPower(RenderWindow * wnd)
 {
 	int effectAlpha = 32;
 
-	RectangleShape rs(Vector2f(1920.f, 500.f));
-	rs.setPosition(0.f, GameDisplay::instance->getSize().y / 2);
+	RectangleShape rs(Vector2f(1920.f, 1080.f));
 	rs.setOrigin(0.f, 250.f);
 	rs.setFillColor(Color(164, 241, 249, effectAlpha));
 	wnd->draw(rs);